+++ lib/runners/processor/flake8.rb
@@ -6,10 +6,10 @@ module Runners
       extend Schema::ConfigTypes
 
       # @type self: SchemaClass
-      let :config, base(
+      let :config, python(
         target: target,
         config: string?,
-        plugins: one_or_more_strings?,
+        plugins: dependencies, # alias
         parallel: boolean?,
       )
     end
@@ -23,20 +23,25 @@ module Runners
     def self.config_example
       <<~'YAML'
         root_dir: project/
+        dependencies:
+          - flake8-bugbear
+          - flake8-builtins==1.4.1
+          - git+https://github.com/PyCQA/flake8-import-order.git@51e16f33065512afa1a85a20b2c2d3be768f78ea
+          - { name: "flake8-docstrings", version: "==1.6.0" }
         target: src/
         config: config/.flake8
-        plugins:
-          - flake8-bandit
-          - flake8-builtins==1.4.1
-          - flake8-docstrings>=1.4.0
-          - git+https://github.com/PyCQA/flake8-import-order.git@51e16f33065512afa1a85a20b2c2d3be768f78ea
         parallel: false
       YAML
     end
 
     def setup
+      begin
+        pip_install Array(config_linter[:dependencies] || config_linter[:plugins])
+      rescue UserError => exn
+        return Results::Failure.new(guid: guid, message: exn.message)
+      end
+
       prepare_config
-      prepare_plugins
       yield
     end
 
@@ -70,10 +75,6 @@ module Runners
           File.delete DEFAULT_CONFIG_PATH
         end
       end
-    end
-
-    def prepare_plugins
-      pip_install(*Array(config_linter[:plugins]))
     end
 
     def parse_result(output)