+++ test/smokes/flake8/option_dependencies/foo.py
@@ -1,0 +1,7 @@
+import sys
+import math
+
+
+def add(a, list=[]):
+    """Add."""
+    return a + list